[setup] Create DRY `EuiCollapsibleNavSubItems` component
- split it out from `EuiCollapsibleNavAccordion`
