+++ src/components/collapsible_nav_beta/collapsible_nav_item/collapsible_nav_accordion.tsx
@@ -11,7 +11,6 @@ import React, {
   ReactNode,
   MouseEvent,
   useCallback,
-  useMemo,
 } from 'react';
 import classNames from 'classnames';
 
@@ -19,7 +18,7 @@ import { useEuiTheme, useGeneratedHtmlId } from '../../../services';
 import { EuiAccordion } from '../../accordion';
 
 import {
-  EuiCollapsibleNavSubItem,
+  EuiCollapsibleNavSubItems,
   EuiCollapsibleNavSubItemProps,
   _SharedEuiCollapsibleNavItemProps,
   _EuiCollapsibleNavItemDisplayProps,
@@ -70,9 +69,6 @@ export const EuiCollapsibleNavAccordion: FunctionComponent<
     accordionProps?.css,
   ];
 
-  /**
-   * Title / accordion trigger
-   */
   const isTitleInteractive = !!(href || linkProps?.onClick);
 
   // Stop propagation on the title so that the accordion toggle doesn't occur on click
@@ -83,40 +79,6 @@ export const EuiCollapsibleNavAccordion: FunctionComponent<
       linkProps?.onClick?.(e);
     },
     [linkProps?.onClick] // eslint-disable-line react-hooks/exhaustive-deps
-  );
-
-  /**
-   * Child items
-   */
-  const itemsHaveIcons = useMemo(
-    () => items.some((item) => !!item.icon),
-    [items]
-  );
-
-  const childrenCssStyles = [
-    styles.children.euiCollapsibleNavAccordion__children,
-    isSubItem ? styles.children.isSubItem : styles.children.isTopItem,
-  ];
-
-  const children = (
-    <div
-      css={childrenCssStyles}
-      className="euiCollapsibleNavAccordion__children"
-    >
-      {items.map((item, index) => {
-        // If any of the sub items have an icon, default to an
-        // icon of `empty` so that all text lines up vertically
-        if (!item.renderItem && itemsHaveIcons && !item.icon) {
-          item.icon = 'empty';
-        }
-        return (
-          // This is an intentional circular dependency between the accordion & parent item display.
-          // EuiSideNavItem is purposely recursive to support any amount of nested sub items,
-          // and split up into separate files/components for better dev readability
-          <EuiCollapsibleNavSubItem key={index} {...item} />
-        );
-      })}
-    </div>
   );
 
   return (
@@ -150,7 +112,11 @@ export const EuiCollapsibleNavAccordion: FunctionComponent<
         ],
       }}
     >
-      {children}
+      <EuiCollapsibleNavSubItems
+        items={items}
+        isSubItem={isSubItem}
+        className="euiCollapsibleNavAccordion__children"
+      />
     </EuiAccordion>
   );
 };
